requirements_test.txt
# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0

# Project dependencies
icmplib>=3.0.0
netifaces>=0.11.0
